make tests generic
