+++ frost-ristretto255/Cargo.toml
@@ -21,7 +21,7 @@ features = ["nightly"]
 byteorder = "1.4"
 curve25519-dalek = { version = "4.0.0-pre.1", features = ["serde"] }
 digest = "0.10"
-frost-core = { path = "../frost-core" }
+frost-core = { path = "../frost-core", features = ["test-impl"] }
 hex = { version = "0.4.3", features = ["serde"] }
 rand_core = "0.6"
 serde = { version = "1", optional = true, features = ["derive"] }
